requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
importlib-metadata==4.6.4
Markdown==3.3.4
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0